Extract normalizeURL into a separate function.
